Search using AJAX - fix bug of Integer.parseINT()..
please guys, don't use parseInt for Long String and double

+++ src/main/java/edu/unsw/cse/comp9323/group1/controllers/CourseDetailController.java
@@ -55,7 +55,7 @@ public class CourseDetailController {
 	 
 		RatingDAO ratingDAO = new RatingDAO();
 		 
-		int overallRatingCourse = ratingDAO.getOverAllRating(courseName);
+		double overallRatingCourse = ratingDAO.getOverAllRating(courseName);
 		HashMap <String, Integer> userRatingsCourse = new HashMap <String, Integer>();
 		userRatingsCourse = ratingDAO.getUserRating(studentId, courseName);
